+++ example-apps/oauth2/middleware.js
@@ -1,0 +1,13 @@
+'use strict';
+
+// This function runs before every outbound request. You can have as many as you
+// need. They'll need to each be registered in your index.js file.
+const includeBearerToken = (request, z, bundle) => {
+  if (bundle.authData.access_token) {
+    request.headers.Authorization = `Bearer ${bundle.authData.access_token}`;
+  }
+
+  return request;
+};
+
+module.exports = { befores: [includeBearerToken], afters: [] };