+++ packages/cli/src/oclif/commands/canary/create.js
@@ -8,6 +8,9 @@ class CanaryCreateCommand extends ZapierBaseCommand {
     const { versionFrom, versionTo } = this.args;
     const percent = this.flags.percent;
     const duration = this.flags.duration;
+    const user = this.flags.user;
+    const accountId = this.flags['account-id'];
+    const forceIncludeAll = this.flags['force-include-all'];
 
     this.validateVersions(versionFrom, versionTo);
     this.validatePercent(percent);
@@ -25,13 +28,34 @@ If you would like to stop this canary now, run \`zapier canary:delete ${existing
       return;
     }
 
-    this.startSpinner(`Creating canary deployment
+    let createCanaryMessage = `Creating canary deployment
     - From version: ${versionFrom}
     - To version: ${versionTo}
-    - Traffic amount: ${percent}%
-    - Duration: ${duration} seconds`);
+    - Percentage: ${percent}%
+    - Duration: ${duration} seconds`;
 
-    await createCanary(versionFrom, versionTo, percent, duration);
+    const body = {
+      percent,
+      duration,
+    };
+
+    if (user) {
+      body.user = user;
+      createCanaryMessage += `\n    - User: ${user}`;
+    }
+
+    if (accountId) {
+      body.account_id = parseInt(accountId);
+      createCanaryMessage += `\n    - Account ID: ${accountId}`;
+    }
+
+    if (forceIncludeAll) {
+      body.force_include_all = true;
+      createCanaryMessage += `\n    - Force Include All: true`;
+    }
+
+    this.startSpinner(createCanaryMessage);
+    await createCanary(versionFrom, versionTo, body);
 
     this.stopSpinner();
     this.log('Canary deployment created successfully.');
@@ -71,6 +95,21 @@ CanaryCreateCommand.flags = buildFlags({
       description: 'Duration of the canary in seconds',
       required: true,
     }),
+    user: Flags.string({
+      char: 'u',
+      description:
+        'Canary this user (email) across all accounts, unless `account-id` is specified.',
+    }),
+    'account-id': Flags.string({
+      char: 'a',
+      description:
+        'The account ID to target. If user is specified, only canary the user within this account. If user is not specified, then this argument is only permitted for Zapier staff.',
+    }),
+    'force-include-all': Flags.boolean({
+      char: 'f',
+      description:
+        'Overrides any default filters the canary system imposes. This argument is only permitted for Zapier staff.',
+    }),
   },
 });
 
@@ -89,13 +128,21 @@ CanaryCreateCommand.description = `Create a new canary deployment, diverting a s
 
 Only one canary can be active at the same time. You can run \`zapier canary:list\` to check. If you would like to create a new canary with different parameters, you can wait for the canary to finish, or delete it using \`zapier canary:delete a.b.c x.y.z\`.
 
+To canary traffic for a specific user, use the --user flag.
+
+To canary traffic for an entire account, use the --account-id. Note: this scenario is only permitted for Zapier staff.
+
+To canary traffic for a specific user within a specific account, use both --user and --account-id flags.
+
 Note: this is similar to \`zapier migrate\` but different in that this is temporary and will "revert" the changes once the specified duration is expired.
 
 **Only use this command to canary traffic between non-breaking versions!**`;
 
 CanaryCreateCommand.examples = [
-  'zapier canary:create 1.0.0 1.1.0 -p 25 -d 720',
-  'zapier canary:create 2.0.0 2.1.0 --percent 50 --duration 300',
+  'zapier canary:create 1.0.0 1.1.0 -p 10 -d 3600',
+  'zapier canary:create 2.0.0 2.1.0 --percent 25 --duration 1800 --user user@example.com',
+  'zapier canary:create 2.0.0 2.1.0 -p 15 -d 7200 -a 12345 -u user@example.com',
+  'zapier canary:create 2.0.0 2.1.0 -p 15 -d 7200 -a 12345',
 ];
 CanaryCreateCommand.skipValidInstallCheck = true;