+++ packages/cli/src/oclif/commands/canary/list.js
@@ -12,6 +12,8 @@ class CanaryListCommand extends ZapierBaseCommand {
       to_version: c.to_version,
       percent: c.percent,
       seconds_remaining: c.until_timestamp - Math.floor(Date.now() / 1000),
+      user: c.user,
+      account_id: c.account_id,
     }));
 
     this.log(bold('Active Canaries') + '\n');
@@ -22,6 +24,8 @@ class CanaryListCommand extends ZapierBaseCommand {
         ['To Version', 'to_version'],
         ['Traffic Amount', 'percent'],
         ['Seconds Remaining', 'seconds_remaining'],
+        ['User', 'user'],
+        ['Account ID', 'account_id'],
       ],
       emptyMessage: grey(`No active canary deployments found.`),
     });