+++ packages/core/src/index.js
@@ -2,6 +2,7 @@
 
 const createLambdaHandler = require('./tools/create-lambda-handler');
 const createAppTester = require('./tools/create-app-tester');
+const { consoleProxy } = require('./tools/console-singleton');
 
 let _integrationTestHandler;
 const integrationTestHandler = (event, context, callback) => {
@@ -15,5 +16,6 @@ module.exports = {
   createAppHandler: createLambdaHandler,
   createAppTester,
   integrationTestHandler,
+  console: consoleProxy,
   ...require('./typeHelpers'),
 };